backend/requirements.txt
# AI B2B Proposal Generator — Backend Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# AI provider
google-generativeai==0.8.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1         # For FastAPI TestClient async support

# Optional: type stubs for better IDE support
# types-jsonschema==4.23.0.1
